# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CentroAbsen API",
    "version": "1.0.0",
    "description": "API Documentation for CentroAbsen HRIS."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local Development"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login Karyawan/User",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "kry001@centroabsen.local" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
...
    "/attendances/{id}/photo": {
      "get": {
        "summary": "Ambil Foto Presensi Terotorisasi (Private)",
        "tags": ["Attendance"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "example": "att-uuid-001" }
          }
        ],
        "responses": {
          "200": { "description": "Stream foto JPEG privat" },
          "401": { "description": "Tidak terotorisasi" },
          "403": { "description": "Akses ditolak" },
          "404": { "description": "Foto tidak ditemukan" }
        }
      }
    },
...
  }
}